refactor: 코드 재작성
