+++ src/main/java/com/dev/voucherproject/VoucherProjectApplication.java
@@ -6,6 +6,7 @@ import org.springframework.boot.SpringApplication;
 import org.springframework.boot.autoconfigure.SpringBootApplication;
 import org.springframework.context.ConfigurableApplicationContext;
 
+
 @SpringBootApplication
 public class VoucherProjectApplication {
     public static void main(String[] args) {
