--- storage/voucher/voucher.csv
@@ -1,2 +1,0 @@
-PERCENT_DISCOUNT_VOUCHER,20,fddf0761-970d-489a-98f4-4ac7f834dc7a
-FIXED_AMOUNT_VOUCHER,3000,edec47ea-c0cb-4afb-ab77-1515f964ddd8
